修改数据权限

+++ erp-model/erp-model-plm/src/main/java/com/erp/model/plm/dto/ProductSaleDTO.java
@@ -30,7 +30,7 @@ public class ProductSaleDTO implements Serializable {
     /**
      * 年目标销售量
      */
-    private Integer yearSaleQty;
+    private Long yearSaleQty;
 
     /**
      * 年目标销售额
@@ -40,7 +40,7 @@ public class ProductSaleDTO implements Serializable {
     /**
      * 月目标销售量
      */
-    private Integer monthSaleQty;
+    private Long monthSaleQty;
 
     /**
      * 月目标销售额
